updated bitcoin to work with new server pattern
