+++ src/api/search_api.rs
@@ -1,0 +1,13 @@
+//! rosetta search api implementation for bitcoin using mentat
+
+use super::*;
+
+/// rosetta data routes for bitcoin
+#[derive(Clone, Default)]
+pub struct BitcoinSearchApi;
+
+#[async_trait]
+impl SearchApiRouter for BitcoinSearchApi {}
+
+#[async_trait]
+impl SearchApi for BitcoinSearchApi {}
