--- src/indexer_api.rs
@@ -1,16 +1,0 @@
-//! rosetta indexer api implementation for bitcoind using mentat
-
-use mentat_server::{
-    api::{CallerIndexerApi, IndexerApi},
-    axum::async_trait,
-};
-
-/// rosetta indexer routes for bitcoind
-#[derive(Clone, Default)]
-pub struct BitcoinIndexerApi;
-
-#[async_trait]
-impl CallerIndexerApi for BitcoinIndexerApi {}
-
-#[async_trait]
-impl IndexerApi for BitcoinIndexerApi {}
