+++ src/responses/response.rs
@@ -12,7 +12,7 @@ use mentat_types::Result;
 use super::ErrorResponse;
 use crate::request::BitcoinJrpc;
 
-/// the response structure for bitcoind
+/// the response structure for bitcoin
 #[derive(Clone, Debug, Deserialize)]
 #[serde(crate = "mentat_server::serde")]
 pub struct Response<R> {
